requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.9
django-filter>=23.3
django-cors-headers>=4.3.0
python-dotenv>=1.0.0
python-dateutil>=2.8.2
django-environ>=0.11.0
requests>=2.31.0
stripe>=7.0.0
gunicorn
